Format is now an enum
Despite the looks, more than two values are allowed per option.
For example there could be a third width format:
```
// Width          XX
ThridOp = 0b00'00'11,
```
Each entry must be unique for bitmask lookup to work, so none of them are zero. (Well one of them can be zero but that would be confusing).